refactor: client-side condition management
Move Condition loading to react-query on the client side.
This means we no longer need to load conditions on every page and can
instead rely on the component to fetch (potentially cached) data
client-side as needed. There's a brief window after loading the page
with uncached data that can result in the (?) cursor before conditions
finish loading, but this seems an acceptable tradeoff.